Massive code changes to replace the use of "integer" and "double" by the single type "number"

+++ src/action/fixtures/simple_email_action_df.sql
@@ -27,8 +27,8 @@ SET default_with_oids = false;
 CREATE TABLE "__ONTASK_WORKFLOW_TABLE_5" (
     age double precision,
     email text,
-    sid bigint,
-    "EmailRead_1" bigint,
+    sid double precision,
+    "EmailRead_1" double precision,
     another text,
     name text,
     one text,
